DePatch 1.18.20
* Changes to grid defender function.
If grid will hit voxel on high speed, it's will not be forced to stop, and will not be pushed to random vector.
Grid will get some damage, to section that hit the voxel.

From many tests this seems to be the best way to handle the voxel collision without chance that ship will go under voxel, or trigger crash in game voxel code.
if ship travel on low speed, there will be no damage if voxel damage set to 0.

* Cleaned not needed code for voxel that was not used.