DePatch 1.18.21
* Added new grid protection system.

if grid is in planet with gravity! and below altitude of 150m, and grid speed is above 70, and grid is above 500 blocks, and it's dont have 4 wheels, and it's has more than 6 thrustes!
THEN if it's hit voxels, it's will be stop, converted to static, and player will be ejected from cockpit and grid will be fixshipped, to unlock option to switch it back to ship from static.

Grid will not have any damage, and only effect on sim is when ship is fixed.. if it's huge, then some small drop detected, if it's around 5K blocks, then almost 0 effect on sim.

This allows Rovers with wheels to be ignored from convertion to static if they have fun driving fast on planet.

There is a check before converting ship to static.
1, it's has to be in gravity above 0.2 and below 20
2. it's has to travel above 70ms or set speed.
3. it's has to have less than 4 wheels with suspension (rover detector)
4. it's has to have 6 or more thrusters! (again rover detector)
5. it's has to be above 500 blocks.
6. it's can be LARGE or SMALL grid.
7. it's has to hit voxels at altitude less than 150mb
8. option has to be enabled.
When converting to static:
all projectors will be OFF
All players will be ejected from pilot sits,
Grid will be stop to 0 ms
grid will be fixship to allow converting back to ship.

* Added new cargo hack fix, to prevent player from creating 1000+ objects in 1 cargo then grinding it and crash server. why? becouse they can.
now they can go to dark place. :)