Build 1.18.66
* Small code cleanup changes.
* Added new Option to Prevent Enemy Player to place ANY block on other player grid! Globably!
Code check if Player is friendly to grid, and check if grid is NPC.
Then if NPC allow placing blocks, if enemy player then deny.
Also ADMIN is allowed to place blocks on any grid.
Grid without Owner is also excluded from this protection.
This option is OFF by default. and checkbox is in Defender Tab.