Take control on Game SAVE code. to fix issues.
* We had problems with game save crashing torch without any log.
i have found that since we have large save .Net GC is responsible for cleaning cached buffer after each save.
So when it's time to do it's task, it's was hitting some bug and torch was closed by force by windows.
To prevent this issue, i have added buffer flush in SerializeXML code,

And added feature to not save large XML each save, but do it once in 25min, this reduced the save size, and save time by 480% from 29sec to 6sec
game can load with the sbcb5 only and XML needed for modifications only.

This code is not for ALL, it's was added to deal with issue for our system.

We had no single crash on  save since this code is active.